+++ Tests/FlareTests/UnitTests/TestHelpers/Stubs/StoreTransactionStub.swift
@@ -54,4 +54,16 @@ final class StoreTransactionStub: IStoreTransaction {
     var environment: StoreEnvironment? {
         stubbedEnvironment
     }
+    
+    var stubbedPrice: Decimal!
+    
+    var price: Decimal? {
+        stubbedPrice
+    }
+    
+    var stubbedCurrency: String!
+    
+    var currency: String? {
+        stubbedCurrency
+    }
 }